test(todo-app): add local e2e tests
Add e2e tests for todo-app that use Appium.

WIP: building and integration tests work locally

WIP: update package.json

WIP: (hack) require bundler late

The bundler depends on react-native's packager. This repo depends on react-native-cli, which assumes any directory with node_modules/react-native is a react-native xcode project. This means the bundler has to depend on the react packager WITHIN the xcode project. It needs to be required late for this reason. It's a hack.